% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_message}
\alias{log_message}
\alias{init_log_file}
\alias{delete_log_file}
\title{Log a message to a file}
\usage{
log_message(
  log_file,
  message,
  context = NULL,
  type = c("INFO", "ACTION", "WARNING", "ERROR"),
  print_level = 1
)

init_log_file(
  usensys_dir,
  log_file = "log.txt",
  case_id = NULL,
  replace = FALSE
)

delete_log_file(
  usensys_dir = NULL,
  log_file = file.path(usensys_dir, "log.txt")
)
}
\arguments{
\item{log_file}{character path to the log file}

\item{message}{character message to log}

\item{context}{optional additional context to log}

\item{type}{character log type, one of 'INFO', 'WARNING', 'ACTION', 'ERROR'}
}
\description{
Log a message to a file
}
\section{Functions}{
\itemize{
\item \code{init_log_file()}: Initialize a log file

\item \code{delete_log_file()}: Delete a log file

}}
