% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mip_build_carbon_cap}
\alias{mip_build_carbon_cap}
\title{Build carbon cap policy from MIP scenario data}
\usage{
mip_build_carbon_cap(usensys_dir, CO2_price_cap = NULL, print_level = 1)
}
\arguments{
\item{print_level}{}
}
\description{
Build carbon cap policy from MIP scenario data
}
