% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mip_build_conv_techs}
\alias{mip_build_conv_techs}
\title{Build conventional technologies from MIP scenario data}
\usage{
mip_build_conv_techs(usensys_dir, group_by_type = FALSE, print_level = 1)
}
\arguments{
\item{print_level}{}
}
\description{
Build conventional technologies from MIP scenario data
}
