% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mip_build_dataset}
\alias{mip_build_dataset}
\title{Build MIP-USENSYS case dataset from GenX inputs}
\usage{
mip_build_dataset(genx_dir, usensys_dir = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Build MIP-USENSYS case dataset from GenX inputs
}
