% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mip_build_demand}
\alias{mip_build_demand}
\title{Build demand from MIP scenario data}
\usage{
mip_build_demand(usensys_dir, print_level = 1)
}
\arguments{
\item{print_level}{}
}
\description{
Build demand from MIP scenario data
}
