% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mip_build_energy_share_requirements}
\alias{mip_build_energy_share_requirements}
\title{Build energy share policy from MIP scenario data}
\usage{
mip_build_energy_share_requirements(usensys_dir, print_level = 1)
}
\arguments{
\item{print_level}{}
}
\description{
Build energy share policy from MIP scenario data
}
