% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mip_build_horizons}
\alias{mip_build_horizons}
\title{Build horizons from MIP scenario data}
\usage{
mip_build_horizons(usensys_dir, print_level = 1)
}
\arguments{
\item{print_level}{}
}
\description{
Build horizons from MIP scenario data
}
