% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mip_build_lostload_penalty}
\alias{mip_build_lostload_penalty}
\title{Build lost load penalty policy from MIP scenario data}
\usage{
mip_build_lostload_penalty(usensys_dir, print_level = 1)
}
\arguments{
\item{print_level}{}
}
\description{
Build lost load penalty policy from MIP scenario data
}
