% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mip_build_model}
\alias{mip_build_model}
\title{Build the MIP-USENSYS model}
\usage{
mip_build_model(usensys_dir, print_level = 1)
}
\arguments{
\item{usensys_dir}{character path to the 'usensys' directory with the
MIP scenario data}

\item{print_level}{}
}
\description{
Build the MIP-USENSYS model
}
