% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mip_build_storage}
\alias{mip_build_storage}
\title{Build storage technologies from MIP scenario data}
\usage{
mip_build_storage(usensys_dir, print_level = 1)
}
\arguments{
\item{print_level}{}
}
\description{
Build storage technologies from MIP scenario data
}
