% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mip_build_supply}
\alias{mip_build_supply}
\title{Build primary supply from MIP scenario data}
\usage{
mip_build_supply(usensys_dir, print_level = 1)
}
\arguments{
\item{print_level}{}
}
\description{
Build primary supply from MIP scenario data
}
