% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mip_build_vre_techs}
\alias{mip_build_vre_techs}
\title{Build VRE technologies from MIP scenario data}
\usage{
mip_build_vre_techs(usensys_dir, print_level = 1)
}
\arguments{
\item{print_level}{}
}
\description{
Build VRE technologies from MIP scenario data
}
