% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mip_build_weather}
\alias{mip_build_weather}
\title{Build weather factors from MIP scenario data}
\usage{
mip_build_weather(usensys_dir, print_level = 1)
}
\arguments{
\item{print_level}{}
}
\description{
Build weather factors from MIP scenario data
}
