% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mip_capacity_factors_mean}
\alias{mip_capacity_factors_mean}
\title{Mean capacity factors for each Resource by year.}
\usage{
mip_capacity_factors_mean(x, na_rm = TRUE, print_level = 1)
}
\arguments{
\item{x, }{a data.table with capacity factors data, or a path to
USENSYS MIP-scenario directory,}

\item{na_rm}{logical indicating if NA values should be removed.}

\item{print_level}{integer indicating the level of verbosity (0 = silent, 1 = print integer columns)}
}
\value{
data.table with mean capacity factors for each Resource by year.
}
\description{
Mean capacity factors for each Resource by year.
}
