% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mip_capacity_factors_na}
\alias{mip_capacity_factors_na}
\title{Find number of NA values in the capacity factors data.}
\usage{
mip_capacity_factors_na(x, print_level = 1)
}
\description{
Find number of NA values in the capacity factors data.
}
