% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mip_capacity_factors_window}
\alias{mip_capacity_factors_window}
\title{Capacity factors data "window" in the data.}
\usage{
mip_capacity_factors_window(x, print_level = 1)
}
\description{
Capacity factors data "window" in the data.
}
