% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mip_check_upper_bounds}
\alias{mip_check_upper_bounds}
\title{Check and correct declining upper bounds for capacity expansion.}
\usage{
mip_check_upper_bounds(
  gen,
  correct_errors = TRUE,
  warn_timer = 5,
  print_level = 1
)
}
\arguments{
\item{gen}{data.table with generators data.}
}
\value{
data.table with corrected upper bounds.
}
\description{
The function checks if the upper bounds are not declining over the years.
Declining upper bounds can cause infeasibility if the resource has
been already built and the upper bound is lower than the existing capacity.
Only technologies available for new build are checked.
The function is called from \code{mip_import_generators} and also can be used
separately.
}
