% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mip_import_capacity_factors}
\alias{mip_import_capacity_factors}
\title{Process capacity factors}
\usage{
mip_import_capacity_factors(
  usensys_dir,
  yaml_file = NULL,
  VER_name_mask = "wind|solar|photovol|utilitypv|hydro|distributed",
  NOT_VER_name_mask =
    "heat|duty|nuclear|coal|gas|oil|biomass|waste|petrol|hydrogen|storage|batter|geothermal",
  print_level = 1
)
}
\arguments{
\item{usensys_dir}{character path to the USENSYS scenario directory.}

\item{VER_name_mask}{character regular expressions to identify VRE generators.
Used to identify "weather-dependent" generators, not marked as VRE in
GenX data (mostly hydroelectric).}

\item{NOT_VER_name_mask}{character regular expressions to exclude non-VRE generators
from the VRE list identified by \code{VER_name_mask}.}

\item{print_level}{integer indicating the level of verbosity (0 = silent, 1 = print integer columns)}
}
\description{
Process capacity factors
}
