% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mip_import_carbon_cap}
\alias{mip_import_carbon_cap}
\title{Process carbon CAP data}
\usage{
mip_import_carbon_cap(usensys_dir, print_level = 1)
}
\arguments{
\item{usensys_dir}{character path to the USENSYS scenario directory.}

\item{print_level}{integer indicating the level of verbosity (0 = silent, 1 = print integer columns)}
}
\description{
Ad-hoc function to process carbon cap data. Reads the data from the
imported GenX data (data-raw/CO2_cap) and "guesses" if the cap is
regional or national. Saves the processed data to the USENSYS model
directory (data/CO2_cap). Hardwired to the MIP's 'CO2_cap' file structure.
Changes of number of regional caps or different file structure will require
changes in the function.
}
