% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mip_import_energy_share_requirement}
\alias{mip_import_energy_share_requirement}
\title{Process energy share requirement data}
\usage{
mip_import_energy_share_requirement(usensys_dir, print_level = 1)
}
\arguments{
\item{usensys_dir}{character path to the USENSYS scenario directory.}

\item{print_level}{integer indicating the level of verbosity (0 = silent, 1 = print integer columns)}
}
\description{
The function reads the energy share requirement data from the imported GenX data
(data-raw/energy_share_requirement) and processes it to prepare inputs for the USENSYS model.
#'
}
