% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mip_import_generators}
\alias{mip_import_generators}
\title{Import generators data to USENSYS data}
\usage{
mip_import_generators(
  usensys_dir,
  correct_errors = TRUE,
  warn_timer = 5,
  print_level = 1
)
}
\arguments{
\item{print_level}{}
}
\description{
The function loads imported from GenX generators data and processes it
to prepare inputs for the USENSYS model. The function also reads
the \code{mip_case.yml}. Both files are required to run the function.
Run \code{mip_import_yaml} and \code{mip_import_raw_data} before calling this function.
}
\examples{
\dontrun{

}
}
