% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mip_import_hourly_loads}
\alias{mip_import_hourly_loads}
\title{Process load data}
\usage{
mip_import_hourly_loads(usensys_dir, print_level = 1)
}
\arguments{
\item{usensys_dir}{character path to the USENSYS scenario directory.}

\item{print_level}{integer indicating the level of verbosity (0 = silent, 1 = print integer columns)}
}
\description{
The function reads the load data from the imported GenX data
(data-raw/load_data) and processes it to prepare inputs for the USENSYS model.
}
