% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mip_import_network}
\alias{mip_import_network}
\title{Process network data}
\usage{
mip_import_network(usensys_dir, print_level = 1)
}
\arguments{
\item{usensys_dir}{character path to the USENSYS scenario directory.}

\item{print_level}{integer indicating the level of verbosity (0 = silent, 1 = print integer columns)}
}
\description{
Reads the network data from inmported GenX data (data-raw/network)
and processes it to prepare inputs for the USENSYS model.
}
