% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mip_import_offshore_wind_min_cap}
\alias{mip_import_offshore_wind_min_cap}
\title{Process minimum capacity data for offshore wind}
\usage{
mip_import_offshore_wind_min_cap(usensys_dir, print_level = 1)
}
\arguments{
\item{usensys_dir}{character path to the USENSYS scenario directory.}

\item{print_level}{integer indicating the level of verbosity (0 = silent, 1 = print integer columns)}
}
\description{
The function reads the minimum capacity data from the imported GenX data
(data-raw/Minimum_capacity_requirement) and processes it to prepare inputs
for the USENSYS model.
}
