% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mip_import_raw_data}
\alias{mip_import_raw_data}
\title{Import GenX inputs and save in usensus/data-raw}
\usage{
mip_import_raw_data(
  genx_dir,
  usensys_dir = NULL,
  yaml_file = NULL,
  overwrite = TRUE,
  clean_start = FALSE,
  print_level = 1,
  prompt_timer = 10
)
}
\arguments{
\item{genx_dir}{character path to GenX input CSV-files.}

\item{usensys_dir}{character path to the destination directory for the USENSYS
scenario files. If NULL, the scenario files are stored in the \code{genx_dir}
directory, \code{usensys} sub-directory.}

\item{yaml_file}{file with GenX-to-USENSYS translation and the case/scenario parameters.}

\item{overwrite}{logical, if TRUE, overwrite existing USENSYS scenario files.}

\item{clean_start}{logical, if TRUE, delete existing USENSYS scenario files.}
}
\value{
\code{arrow-parquet} files with USENSYS scenario data, partitioned by year.
}
\description{
Ad-hoc function to read GenX input files and store in
\code{arrow-parquet} format for further use to build energyRt/USENSYS models.
}
\details{
The function uses the model structure and assumptions of MIP project,
and does not take into account parameters and features not used in the project.
}
