% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mip_import_representative_period}
\alias{mip_import_representative_period}
\title{Process representative period data}
\usage{
mip_import_representative_period(usensys_dir, print_level = 1)
}
\arguments{
\item{usensys_dir}{character path to the USENSYS scenario directory.}

\item{print_level}{integer indicating the level of verbosity (0 = silent, 1 = print integer columns)}
}
\description{
Process representative period data
}
