% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{mip_import_yaml}
\alias{mip_import_yaml}
\title{Create a YAML file for the MIP scenario based on GenX input files and a template}
\usage{
mip_import_yaml(
  genx_dir,
  usensys_dir = file.path(genx_dir, "usensys"),
  yaml_file = "mip_case.yml",
  case_id = NULL,
  case_description = NULL,
  print_level = 1,
  update = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{genx_dir}{character: The path to the GenX input folder}

\item{yaml_file}{character: The name of the YAML file to create}

\item{print_level}{numeric: The print level (0: no print, 1: print, 2: verbose)}
}
\description{
The function scans through PowerGenome/GenX yaml files for mip case settings and
creates a new YAML file with the specified parameters for the MIP case/scenario.
}
