% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{mip_load_yaml}
\alias{mip_load_yaml}
\alias{mip_read_yaml}
\title{Load a YAML file for the MIP scenario}
\usage{
mip_load_yaml(usensys_dir, yaml_file = "mip_case.yml")

mip_read_yaml(usensys_dir, yaml_file = "mip_case.yml")
}
\arguments{
\item{usensys_dir}{character: The path to the USENSYS directory}

\item{yaml_file}{character: The name of the YAML file to load}
}
\value{
A list with the YAML content
}
\description{
Load a YAML file for the MIP scenario
}
