% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mip_make_fuel_groups}
\alias{mip_make_fuel_groups}
\title{Create fuel groups mapping table and save to the USENSYS data directory.}
\usage{
mip_make_fuel_groups(
  usensys_dir,
  fuels = NULL,
  save_data = FALSE,
  print_level = 1
)
}
\arguments{
\item{usensys_dir}{character path to the USENSYS scenario directory.}

\item{save_data}{}

\item{print_level}{integer indicating the level of verbosity (0 = silent, 1 = print integer columns)}
}
\description{
Create fuel groups mapping table and save to the USENSYS data directory.
}
