% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mip_make_resource_groups}
\alias{mip_make_resource_groups}
\title{Create resource groups and save to the USENSYS data directory.}
\usage{
mip_make_resource_groups(usensys_dir, gen = NULL, print_level = 1)
}
\arguments{
\item{print_level}{}
}
\description{
Create resource groups and save to the USENSYS data directory.
}
