% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mip_make_zone_region}
\alias{mip_make_zone_region}
\title{Process region names and zone-region mapping.}
\usage{
mip_make_zone_region(usensys_dir, print_level = 1)
}
\arguments{
\item{usensys_dir}{character path to the USENSYS scenario directory.}

\item{print_level}{integer indicating the level of verbosity (0 = silent, 1 = print integer columns)}
}
\description{
Process region names and zone-region mapping.
}
