% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mip_match_cf_process_lifetime}
\alias{mip_match_cf_process_lifetime}
\title{Match capacity factors 'window' in data with process lifetime data.}
\usage{
mip_match_cf_process_lifetime(
  usensys_dir,
  all_resouces = FALSE,
  print_level = 1
)
}
\arguments{
\item{usensys_dir}{character path to the USENSYS scenario directory.}

\item{all_resouces}{logical indicating if all resources should be included.
The default is FALSE, meaning only resources with capacity factors
(both filtered by \code{mip_import_capacity_factors} function) are considered.}

\item{print_level}{}
}
\description{
The function compares the range of years for capacity factors in the
data with the lifetime window for each Resource.
}
