% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{mip_open_yaml}
\alias{mip_open_yaml}
\title{Open a YAML file in the default text editor}
\usage{
mip_open_yaml(usensys_dir, yaml_file = "mip_case.yml")
}
\arguments{
\item{usensys_dir}{character: The path to the USENSYS directory}

\item{yaml_file}{character: The name of the YAML file to open}
}
\description{
Open a YAML file in the default text editor
}
