% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mip_resource_lifetime}
\alias{mip_resource_lifetime}
\title{Calculate lifetime window for each Resource in the data.}
\usage{
mip_resource_lifetime(x, print_level = 1)
}
\arguments{
\item{print_level}{integer indicating the level of verbosity (0 = silent, 1 = print integer columns)}

\item{gen}{data.frame with generators data.}

\item{usensys_dir}{character path to the USENSYS scenario directory.}
}
\value{
data.table with lifetime windows for each Resource.
}
\description{
Calculate lifetime window for each Resource in the data.
}
