% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mip_resource_type}
\alias{mip_resource_type}
\title{Map technology names to technology types.}
\usage{
mip_resource_type(x, groups = 17)
}
\arguments{
\item{x}{character vector with technology names.}

\item{groups}{integer number of groups to use (currently only 17 is supported).}
}
\value{
character vector with technology types.
}
\description{
Map technology names to technology types.
}
\keyword{internal}
