% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mip_suggest_integer_columns}
\alias{mip_suggest_integer_columns}
\title{A utility function to identify integer columns in a data frame}
\usage{
mip_suggest_integer_columns(
  data,
  min_value = -100,
  max_value = 100,
  threshold = 1,
  print_level = 1
)
}
\arguments{
\item{data}{data frame, data table, or arrow query result}

\item{min_value}{minimum value for a column to be considered an integer}

\item{max_value}{maximum value for a column to be considered an integer}

\item{threshold}{minimum proportion of integer values in a column (default 1.0 - all values)}

\item{print_level}{integer indicating the level of verbosity (0 = silent, 1 = print integer columns)}
}
\value{
character vector of column names identified as integers
}
\description{
The function is used to suggest integer data types imported as double
from CSV files. It is used to create an internal list of 'known integer columns'
that are used to force integer data types when reading CSV files from
GenX/PowerGenome files.
}
