% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{mip_write_yaml}
\alias{mip_write_yaml}
\title{Create a YAML file for the MIP scenario}
\usage{
mip_write_yaml(
  usensys_dir = NULL,
  yaml_file = file.path(usensys_dir, "mip_case.yml"),
  title = "MIP case settings for the USENSYS model",
  case_id = "CASE ID-NAME (REQUIRED)",
  case_description = "CASE_DESCRIPTION (OPTIONAL)",
  input_idx = c(1, 2, 3, 4, 5, 6),
  YEARS = c(2027, 2030, 2035, 2040, 2045, 2050),
  group_techs = FALSE,
  group_fuels = FALSE,
  use_short_names = TRUE,
  invcost = FALSE,
  correct_declining_bounds = TRUE,
  drop_ver_no_cf = TRUE,
  drop_cf_no_techs = TRUE,
  drop_zero_capacity_no_build = TRUE,
  vintage = FALSE,
  calendar = "w52_h168",
  subset_weeks = 0,
  SDR = 0.02,
  WACC = 0.05,
  ESR = TRUE,
  lostload = TRUE,
  optimize_storage_duration = TRUE,
  storage_full_year = FALSE,
  early_retirement = FALSE,
  CO2_cap_national = FALSE,
  CO2_cap_regional = TRUE,
  offshore_wind_min_capacity = TRUE,
  backstop = TRUE,
  RAMP = FALSE,
  myopic = TRUE,
  solver = "energyRt::solver_options$gams_gdx_cplex_barrier",
  overwrite = FALSE,
  print_level = 1
)
}
\arguments{
\item{yaml_file}{character: The name of the YAML file to create}

\item{title}{character: The title of the YAML file}

\item{case_id}{character: The name of the case/scenario, required field
in further processing of the YAML file}

\item{case_description}{character: The description of the case/scenario}

\item{input_idx}{numeric: The indices of the input folders to use
(Inputs_p{*} in the GenX Inputs folder)}

\item{YEARS}{numeric: The actual years in the input folders}

\item{group_techs}{logical: If TRUE, removes the zone identifier from the technology name}

\item{group_fuels}{logical: If TRUE, removes the zone identifier from the fuel name}

\item{use_short_names}{logical: If TRUE (recommended), uses short names for technologies and fuels}

\item{invcost}{logical: If TRUE, overnight investment costs will be used
from the GenX input files to calculate annualized costs, if FALSE,
the annualized costs are used directly from the GenX input files (default)}

\item{vintage}{logical: (experimental!) If TRUE, year identifiers are added
to the technology names}

\item{calendar}{character: The calendar settings ("w52_h168" used in all MIP cases)}

\item{subset_weeks}{numeric: The number of weeks to subset (0 for automatic detection)}

\item{SDR}{numeric: The social discount rate (used in perfect foresight scenarios),
default is 0.02 (2\%) in all MIP cases}

\item{WACC}{numeric: The weighted average cost of capital, default is 0.05 (5\%)
in all MIP cases}

\item{ESR}{logical: If TRUE, the Renewable Portfolio Standard (RPS) is used
as a policy, used in the current policy scenarios}

\item{lostload}{numeric: The lost load penalty in $/MWh (default is 500)}

\item{CO2_cap_national}{logical: If TRUE, a national CO2 cap is used}

\item{CO2_cap_regional}{logical: If TRUE, a regional CO2 cap is used, used in
the MIP current policy scenarios}

\item{offshore_wind_min_capacity}{logical: If TRUE,
the offshore wind minimum capacity requirement constraint added}

\item{backstop}{numeric: The backstop price in $/CO2-ton (default is 200)}

\item{RAMP}{logical: (optional) If TRUE, the ramping constraints are used}

\item{myopic}{logical: If TRUE, the myopic solution is used, if FALSE, the perfect}

\item{solver}{character: The solver settings in energyRt format
(default is "energyRt::solver_options$gams_gdx_cplex_barrier")}
}
\description{
This function creates a YAML file with the specified parameters for the MIP case/scenario.
}
\examples{
\dontrun{
mip_write_yaml("mip_case.yaml")
}
}
