% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{update_yaml}
\alias{update_yaml}
\title{Update multiple parameters in a YAML file}
\usage{
update_yaml(
  yaml_file,
  output_file = yaml_file,
  ...,
  print_level = 1,
  overwrite = TRUE
)
}
\arguments{
\item{yaml_file}{character: The path to the YAML file}

\item{output_file}{character: The path to the output file (default is the input file)}

\item{...}{named parameters to update}

\item{print_level}{numeric: The print level (0: no print, 1: print, 2: verbose)}

\item{overwrite}{logical: If TRUE, the output file will be overwritten}
}
\description{
Update multiple parameters in a YAML file
}
